requirements.txt
streamlit==1.51.0
PyPDF2==3.0.1
requests==2.31.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
loguru==0.7.2
groq
reportlab